requirements.txt
flask
flask-cors
langchain
python-dotenv
chromadb
openai
tiktoken
google-cloud-storage
requests